Credit-Card-Fraud-Detection-using-Google-Cloud/requirements.txt
apache-beam[gcp]
google-cloud-pubsub
google-cloud-bigquery
google-cloud-firestore
google-cloud-aiplatform
oauth2client
google-cloud-secret-manager

